face_detect/requirements.txt
# 人脸检测与活体检测GUI应用依赖包
# Face Detection and Liveness Detection GUI Application Dependencies

# 核心依赖 / Core Dependencies
opencv-python>=4.8.0
mediapipe>=0.10.0
Pillow>=10.0.0
pydantic>=2.0.0
numpy>=1.24.0

# 活体检测额外依赖 / Additional Dependencies for Liveness Detection
scipy>=1.10.0       # 用于科学计算 / For scientific computing
sklearn>=1.3.0      # 用于机器学习算法 / For machine learning algorithms

# GUI依赖 / GUI Dependencies
# tkinter 是Python标准库的一部分，通常不需要单独安装
# tkinter is part of Python standard library, usually no need to install separately

# 可选依赖 / Optional Dependencies
# matplotlib>=3.7.0  # 用于数据可视化 / For data visualization
# seaborn>=0.12.0    # 用于统计图表 / For statistical plots